+++ test/live-preview/next-app/app/_api/fetchPage.ts
@@ -1,8 +1,6 @@
 import { Page } from '@/payload-types'
 
-export const PAYLOAD_SERVER_URL = 'http://localhost:3000'
-
-export const getPage = async (slug: string): Promise<Page> => {
+export const fetchPage = async (slug: string): Promise<Page> => {
   return await fetch(`http://localhost:3000/api/pages?where[slug][equals]=${slug}`, {
     method: 'GET',
     cache: 'no-store',
