+++ test/live-preview/next-app/app/page.tsx
@@ -1,7 +1,7 @@
-import { getPage } from './api'
 import { Page } from './page.client'
 import { notFound } from 'next/navigation'
 import type { Metadata, ResolvingMetadata } from 'next'
+import { fetchPage } from './_api/fetchPage'
 
 type Props = {
   params: { id: string }
@@ -12,16 +12,16 @@ export async function generateMetadata(
   { params, searchParams }: Props,
   parent: ResolvingMetadata,
 ): Promise<Metadata> {
-  const page = await getPage('home')
+  const page = await fetchPage('home')
 
   return {
-    title: page.metaTitle || page.title,
-    description: page.metaDescription,
+    title: page.meta?.title || page.title,
+    description: page.meta?.description,
   }
 }
 
 export default async function Home() {
-  const page = await getPage('home')
+  const page = await fetchPage('home')
 
   if (!page) {
     notFound()
